added 2nd translate in draw to reset origin
Hey Dan!
This change seems to fix issue #28 (Starfield p5.js version does not produce output as p5).  Translates accumulate? But shouldn't translates start fresh at every loop, accumulating only within the draw loop? I still get tripped up on that.
 
The Bryn Mawr version of the starfield also has a translate in the draw loop, but they've a secondary translate along with a push and pop matrix in the Star object's draw function. Not that I know whether it works for their version, I haven't run their code, just noticed the difference.  Something seems odd.

Okay, verified in Processing 3 that translate resets itself at every loop, whereas p5.js does not.  If in p5 you set translate in setup(), the translate would remain through the draw loop unchanged, but also would not accumulate.  Perhaps this is a fundamental difference with p5? and the real fix for issue #28 is to place translate in setup and remove it from draw()?  Or is this a known issue?

Also, I'm grateful you ditched the challenge countdown.  The 25 seconds spent describing the star's location as a ratio from initial location to the edge was like David Foster Wallace explaining all of Infinite Jest with a single sentence hidden somewhere in chapter 2.   I tried working that idea into standard form (acceleration -> velocity -> location) but made the mistake of using location inside the map function as such: 
      velocity.x = map(location.x / z, 0, 1, 0, width)
      location.add(velocity);